Optimizacion Pregunta Controller

+++ app/Http/Controllers/PreguntaController.php
@@ -11,48 +11,46 @@ use Auth;
 use Redirect;
 use Illuminate\Routing\Route;
 use DB;
-class PreguntaController extends Controller
-{
-    public function __construct(){
-        $this->beforeFilter('@find', ['only' => ['edit', 'update', 'destroy']]);
+class PreguntaController extends Controller{
+  public function __construct(){
+    $this->beforeFilter('@find', ['only' => ['edit', 'update', 'destroy']]);
+  }
+  public function find(Route $route){
+    $this->pregunta = Pregunta::find($route->getParameter('preguntas'));
+  }
+  public function index(){
+    $preguntas = Pregunta::paginate(5);
+    return view('preguntas.index', compact('preguntas'));
+  }
+  public function create(){
+      return view('preguntas.create');
+  }
+  public function store(PreguntaCreateRequest $request){
+      Pregunta::create($request->all());
+      Session::flash('message', 'Pregunta creada correctamente');
+      return Redirect::to('/preguntas');
+  }
+  public function show($id){
+
+  }
+  public function edit($id){
+    if(isset($id)){
+      return view('preguntas.edit', ['pregunta' => $this->pregunta]);
     }
-    public function find(Route $route){
-        $this->pregunta = Pregunta::find($route->getParameter('preguntas'));
-    }        
-    public function index()
-    {
-        $preguntas = Pregunta::paginate(5);
-        return view('preguntas.index', compact('preguntas'));
+    return response()->json('Acceso denegado');
+  }
+  public function update(PreguntaUpdateRequest $request, $id){
+    $this->pregunta->fill($request->all());
+    $this->pregunta->save();
+    Session::flash('message', 'Pregunta editada correctamente');
+    return Redirect::to('/preguntas');
+  }
+  public function destroy($id){
+    if(isset($id)){
+      $this->pregunta->delete();
+      Session::flash('message', 'Pregunta eliminada correctamente');
+      return Redirect::to('/preguntas');
     }
-    public function create()
-    {
-        return view('preguntas.create');        
-    }
-    public function store(PreguntaCreateRequest $request)
-    {
-        Pregunta::create($request->all());
-        Session::flash('message', 'Pregunta creada correctamente');
-        return Redirect::to('/preguntas');        
-    }
-    public function show($id)
-    {
-                
-    }
-    public function edit($id)
-    {
-        return view('preguntas.edit', ['pregunta' => $this->pregunta]);      
-    }
-    public function update(PreguntaUpdateRequest $request, $id)
-    {
-        $this->pregunta->fill($request->all());
-        $this->pregunta->save();
-        Session::flash('message', 'Pregunta editada correctamente');
-        return Redirect::to('/preguntas');       
-    }
-    public function destroy($id)
-    {
-        $this->pregunta->delete();
-        Session::flash('message', 'Pregunta eliminada correctamente');
-        return Redirect::to('/preguntas');       
-    }
+    return response()->json('Acceso denegado');
+  }
 }
